requirements.txt
python-telegram-bot==20.3
sentence-transformers==2.2.2
scikit-learn==1.4.1
numpy==1.25.2
